requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
aiofiles==24.1.0
    # via gql
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.12.15
    # via gql
aiosignal==1.4.0
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anyio==4.10.0
    # via
    #   gql
    #   httpx
async-timeout==5.0.1
    # via aiohttp
attrs==25.3.0
    # via aiohttp
backoff==2.2.1
    # via gql
botocore==1.40.21
    # via gql
build==1.3.0
    # via pip-tools
certifi==2025.8.3
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.17.1
    # via
    #   cryptography
    #   pynacl
charset-normalizer==3.4.3
    # via requests
click==8.2.1
    # via pip-tools
colorama==0.4.6
    # via
    #   build
    #   click
    #   loguru
cryptography==45.0.6
    # via pyjwt
exceptiongroup==1.3.0
    # via anyio
frozenlist==1.7.0
    # via
    #   aiohttp
    #   aiosignal
gql[all]==4.0.0
    # via -r requirements.in
graphql-core==3.2.6
    # via gql
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via gql
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
importlib-metadata==8.7.0
    # via build
jmespath==1.0.1
    # via botocore
loguru==0.7.3
    # via -r requirements.in
multidict==6.6.4
    # via
    #   aiohttp
    #   yarl
packaging==25.0
    # via build
pip-tools==7.5.0
    # via -r requirements.in
propcache==0.3.2
    # via
    #   aiohttp
    #   yarl
pycparser==2.22
    # via cffi
pydantic==2.11.7
    # via
    #   -r requirements.in
    #   pydantic-settings
pydantic-core==2.33.2
    # via pydantic
pydantic-settings==2.10.1
    # via -r requirements.in
pygithub==2.7.0
    # via -r requirements.in
pyjwt[crypto]==2.10.1
    # via pygithub
pynacl==1.5.0
    # via pygithub
pyproject-hooks==1.2.0
    # via
    #   build
    #   pip-tools
python-dateutil==2.9.0.post0
    # via botocore
python-dotenv==1.1.1
    # via
    #   -r requirements.in
    #   pydantic-settings
requests==2.32.5
    # via
    #   gql
    #   pygithub
    #   requests-toolbelt
requests-toolbelt==1.0.0
    # via gql
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via anyio
tomli==2.2.1
    # via
    #   build
    #   pip-tools
typing-extensions==4.15.0
    # via
    #   aiosignal
    #   anyio
    #   exceptiongroup
    #   multidict
    #   pydantic
    #   pydantic-core
    #   pygithub
    #   typing-inspection
typing-inspection==0.4.1
    # via
    #   pydantic
    #   pydantic-settings
urllib3==2.5.0
    # via
    #   botocore
    #   pygithub
    #   requests
websockets==15.0.1
    # via gql
wheel==0.45.1
    # via pip-tools
win32-setctime==1.2.0
    # via loguru
yarl==1.20.1
    # via
    #   aiohttp
    #   gql
zipp==3.23.0
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
